Got rid of the shared library
