+++ ModernUO.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SerializationGenerator", "P
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SerializationGenerator.Tests", "Projects\SerializationGenerator.Tests\SerializationGenerator.Tests.csproj", "{2BC92375-BB66-4CA5-B39C-36215749FE64}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeGenerationShared", "Projects\CodeGenerationShared\CodeGenerationShared.csproj", "{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SerializationSchemaGenerator", "Projects\SerializationSchemaGenerator\SerializationSchemaGenerator.csproj", "{A30150A3-796C-4C6D-B3E4-B7BEB0021701}"
 EndProject
 Global
@@ -69,12 +67,6 @@ Global
 		{2BC92375-BB66-4CA5-B39C-36215749FE64}.Debug|x64.Build.0 = Debug|x64
 		{2BC92375-BB66-4CA5-B39C-36215749FE64}.Release|x64.ActiveCfg = Release|x64
 		{2BC92375-BB66-4CA5-B39C-36215749FE64}.Release|x64.Build.0 = Release|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Analyze|x64.ActiveCfg = Analyze|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Analyze|x64.Build.0 = Analyze|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Debug|x64.ActiveCfg = Debug|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Debug|x64.Build.0 = Debug|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Release|x64.ActiveCfg = Release|x64
-		{FD81A2D9-F4AE-447E-A13B-C018AD7BD10D}.Release|x64.Build.0 = Release|x64
 		{A30150A3-796C-4C6D-B3E4-B7BEB0021701}.Analyze|x64.ActiveCfg = Analyze|x64
 		{A30150A3-796C-4C6D-B3E4-B7BEB0021701}.Analyze|x64.Build.0 = Analyze|x64
 		{A30150A3-796C-4C6D-B3E4-B7BEB0021701}.Debug|x64.ActiveCfg = Debug|x64
