--- Projects/CodeGenerationShared/SerializableMigration/SerializableMetadata.cs
@@ -1,32 +1,0 @@
-/*************************************************************************
- * ModernUO                                                              *
- * Copyright 2019-2021 - ModernUO Development Team                       *
- * Email: hi@modernuo.com                                                *
- * File: SerializableMigration.cs                                        *
- *                                                                       *
- * This program is free software: you can redistribute it and/or modify  *
- * it under the terms of the GNU General Public License as published by  *
- * the Free Software Foundation, either version 3 of the License, or     *
- * (at your option) any later version.                                   *
- *                                                                       *
- * You should have received a copy of the GNU General Public License     *
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
- *************************************************************************/
-
-using System.Collections.Immutable;
-using System.Text.Json.Serialization;
-
-namespace CodeGeneration
-{
-    public record SerializableMetadata
-    {
-        [JsonPropertyName("version")]
-        public int Version { get; init; }
-
-        [JsonPropertyName("type")]
-        public string Type { get; init; }
-
-        [JsonPropertyName("properties")]
-        public ImmutableArray<SerializableProperty> Properties { get; init; }
-    }
-}
