--- Projects/CodeGenerationShared/SourceGeneration/SourceGeneration.AccessModifier.cs
@@ -1,43 +1,0 @@
-/*************************************************************************
- * ModernUO                                                              *
- * Copyright (C) 2019-2021 - ModernUO Development Team                   *
- * Email: hi@modernuo.com                                                *
- * File: SourceGeneration.AccessModifier.cs                              *
- *                                                                       *
- * This program is free software: you can redistribute it and/or modify  *
- * it under the terms of the GNU General Public License as published by  *
- * the Free Software Foundation, either version 3 of the License, or     *
- * (at your option) any later version.                                   *
- *                                                                       *
- * You should have received a copy of the GNU General Public License     *
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
- *************************************************************************/
-
-namespace CodeGeneration
-{
-    public enum AccessModifier
-    {
-        None,
-        Public,
-        Private,
-        Protected,
-        Internal,
-        ProtectedInternal,
-        PrivateProtected
-    }
-
-    public static partial class SourceGeneration
-    {
-        public static string ToFriendlyString(this AccessModifier modifier) =>
-            modifier switch
-            {
-                AccessModifier.Public => "public",
-                AccessModifier.Private => "private",
-                AccessModifier.Protected => "protected",
-                AccessModifier.Internal => "internal",
-                AccessModifier.ProtectedInternal => "protected internal",
-                AccessModifier.PrivateProtected => "private protected",
-                _ => ""
-            };
-    }
-}
