+++ Projects/SerializationGenerator/SerializableMigration/Rules/SerializableInterfaceMigrationRule.cs
@@ -1,0 +1,72 @@
+/*************************************************************************
+ * ModernUO                                                              *
+ * Copyright 2019-2021 - ModernUO Development Team                       *
+ * Email: hi@modernuo.com                                                *
+ * File: SerializableInterfaceMigrationRule.cs                           *
+ *                                                                       *
+ * This program is free software: you can redistribute it and/or modify  *
+ * it under the terms of the GNU General Public License as published by  *
+ * the Free Software Foundation, either version 3 of the License, or     *
+ * (at your option) any later version.                                   *
+ *                                                                       *
+ * You should have received a copy of the GNU General Public License     *
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
+ *************************************************************************/
+
+using System;
+using System.Collections.Immutable;
+using System.Text;
+using Microsoft.CodeAnalysis;
+using SourceGeneration;
+
+namespace SerializableMigration
+{
+    public class SerializableInterfaceMigrationRule : ISerializableMigrationRule
+    {
+        public string RuleName => nameof(SerializableInterfaceMigrationRule);
+
+        public bool GenerateRuleState(
+            Compilation compilation,
+            ISymbol symbol,
+            ImmutableArray<AttributeData> attributes,
+            ImmutableArray<INamedTypeSymbol> serializableTypes,
+            out string[] ruleArguments
+        )
+        {
+            if (symbol is ITypeSymbol typeSymbol && typeSymbol.HasSerializableInterface(compilation, serializableTypes))
+            {
+                ruleArguments = Array.Empty<string>();
+                return true;
+            }
+
+            ruleArguments = null;
+            return false;
+        }
+
+        public void GenerateDeserializationMethod(StringBuilder source, string indent, SerializableProperty property)
+        {
+            const string expectedRule = nameof(SerializableInterfaceMigrationRule);
+            var ruleName = property.Rule;
+            if (expectedRule != ruleName)
+            {
+                throw new ArgumentException($"Invalid rule applied to property {ruleName}. Expecting {expectedRule}, but received {ruleName}.");
+            }
+
+            var propertyName = property.Name;
+            source.AppendLine($"{indent}{propertyName} = reader.ReadEntity<{property.Type}>();");
+        }
+
+        public void GenerateSerializationMethod(StringBuilder source, string indent, SerializableProperty property)
+        {
+            const string expectedRule = nameof(SerializableInterfaceMigrationRule);
+            var ruleName = property.Rule;
+            if (expectedRule != ruleName)
+            {
+                throw new ArgumentException($"Invalid rule applied to property {ruleName}. Expecting {expectedRule}, but received {ruleName}.");
+            }
+
+            var propertyName = property.Name;
+            source.AppendLine($"{indent}writer.Write({propertyName});");
+        }
+    }
+}
